+++ src/BlockMath.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import createMathComponent from './createMathComponent';
+import createMathComponent from './createMathComponent.jsx';
 
 const BlockMath = ({ html }) => {
   return <div dangerouslySetInnerHTML={{__html: html}} />;
