+++ src/InlineMath.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import createMathComponent from './createMathComponent';
+import createMathComponent from './createMathComponent.jsx';
 
 const InlineMath = ({ html }) => {
   return <span dangerouslySetInnerHTML={{__html: html}} />;
