+++ src/createMathComponent.jsx
